Make tests use the actual version of Likely instead of specifically generated one
I’ve made a stupid mistake: all tests were ran against a customly-generated version of Likely hosted in gh-pages. This made all the CI integration just meaningless. This commit fixes this.

I tried to use `file://` urls initially, but this was resulting in a `TypeError: this.browser.getBrowserForTab is not a function` error, so I had to use a local HTTP server instead.