Add Viber service provider
In Viber sharing works like forwarding, so at least in some versions it's not possible to edit a shared message. Because of that it's not a good idea to use title option (which has a fallback to page title in case it's not set by user). So instead we allow user to specify message with data-attribute and handle it with new option, so that we can pass everything in one query parameter without messing with content of url / title.