- Update viber svg icon path so that it's visually consistent with similar whatsapp icon
For whatsapp we currently use only the inner circle part of the official logo, while viber has the whole logo wrapped in a circle. The example for viber was found on their official brand page, but it kinda assumes using round badges in general (for example, twitter badge as a smaller bird inside a circle) and we don't really do it accross the system. Anyway, now we mainly want those two to look similar, so make the recently added one (viber) use the same approach as what we already had (whatsapp)