Introduce (temporarily) a concept of known data attributes for a service (ones that are going to be used to construct a link when specified on button level)
Version 3.0 is going to have a breaking change where each service will construct a share url based on its own capabilities (in opposite to having a unified method of appending all service data attributes to a link). Until that happens we want to throw a deprecation warning if data attribute is not supported by likely officially.